# Optimized TreeNode.siblings and neighbors